Move Recycler to a separate module
